test/src/requirements.txt
Django==2.2
django-crispy-forms==1.8.1
Pillow==7.0.0
django-storages==1.91
boto3==1.12.42